Remove validate_python_dependencies when create extension
The python dependencies could be installed in the virtual env. However,
the 'pgml.activate_venv()' will be available after the 'CREATE
EXTENSION' statement.

In this situation, the `CREATE EXTENSION` always fails and user doesn't
have a chance to set the venv.

Also, the python dependencies could be removed at any time after the
extension creation. A validation at the beginning is not quite useful.